# Speed up exclusions with picomatch 