--- a/snowpack/src/commands/build.ts
+++ b/snowpack/src/commands/build.ts
@@ -1,7 +1,7 @@
 import {ImportMap, InstallTarget} from 'esinstall';
 import {promises as fs} from 'fs';
 import {fdir} from 'fdir';
-import micromatch from 'micromatch';
+import picomatch from 'picomatch';
 import * as colors from 'kleur/colors';
 import mkdirp from 'mkdirp';
 import path from 'path';
@@ -105,8 +105,9 @@ export async function build(commandOptions: CommandOptions): Promise<SnowpackBui
     const files = (await new fdir().withFullPaths().crawl(mountKey).withPromise()) as string[];
     const excludePrivate = new RegExp(`\\${path.sep}\\.`);
     const excludeGlobs = [...config.exclude, ...config.testOptions.files];
+    const foundExcludeMatch = picomatch(excludeGlobs);
     for (const f of files) {
-      if (micromatch.isMatch(f, excludeGlobs) || excludePrivate.test(f)) {
+      if (foundExcludeMatch(f) || excludePrivate.test(f)) {
         continue;
       }
       const fileUrls = getUrlsForFile(f, config)!;