## Changes

In https://github.com/snowpackjs/snowpack/pull/2876, `fdir` was used to speed up mounting directories in large repos (+50k files) from +15s down to <500ms. However, the use of `micromatch` to [respect exclusions](https://github.com/snowpackjs/snowpack/pull/2876#discussion_r595622596) introduced a regressions where bootups shot up to +60s.

This PR replaces [micromatch](https://github.com/micromatch/micromatch) with [picomatch](https://github.com/micromatch/picomatch) to speed up glob pattern matching for file exclusions. 
 
### Before (+60s)
<img width="538" alt="Screen Shot 2021-03-18 at 2 19 57 AM" src="https://user-images.githubusercontent.com/6130700/111582215-e7dd1c00-8790-11eb-9487-d9ba525fc6da.png">


### After (~900ms)
<img width="509" alt="Screen Shot 2021-03-18 at 2 49 12 PM" src="https://user-images.githubusercontent.com/6130700/111685042-d67e2900-87fd-11eb-8189-b6787fb78103.png">

## Testing

- Build the project: `yarn  run build && yarn run bundle`
- Copy into local project: `cp  ./snowpack/lib/index.js ../../my-project/node_modules/snowpack/lib/index.js`
- Test bootup: `yarn run snowpack dev`