@FredKSchott 2021-03-18
LGTM! I'm going to merge so that we can test this on main a bit more, but given the convo above I think we can go back to not running picomatch at all in the `dev.ts` file, since that's the one place where there's no significant cost for handling ignored files.